rules-api/pyproject.toml
[tool.poetry]
name = "rules-api"
version = "0.1.0"
description = "Rules API for rules project."
authors = ["Bogdan Gulowaty <b.gulowaty@gmail.com>"]
packages = [
    { include = "rules" },
]

[tool.poetry.dependencies]
python = "^3.9"
attrs = "^20.3.0"
toolz = "^0.11.1"
joblib = "^1.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
